Basic Functions Finished

+++ app/src/main/java/com/aritxonly/deadliner/MainActivity.kt
@@ -1,30 +1,34 @@
 package com.aritxonly.deadliner
+import android.content.Context
 import android.content.Intent
-import android.graphics.Color
+import android.content.SharedPreferences
+import android.graphics.*
 import android.os.Build
 import android.os.Bundle
-import android.os.Debug
 import android.os.Handler
 import android.os.Looper
+import android.os.VibrationEffect
+import android.os.Vibrator
 import android.util.Log
 import android.util.TypedValue
 import android.view.View
-import android.view.Window
 import android.view.WindowInsetsController
 import android.view.WindowManager
-import android.widget.Button
 import android.widget.ImageButton
+import android.widget.TextView
 import android.widget.Toast
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.google.android.material.color.DynamicColors
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.android.material.floatingactionbutton.FloatingActionButton
-import java.time.LocalDateTime
+import kotlinx.coroutines.*
 
 class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
 
@@ -35,6 +39,10 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
     private val itemList = mutableListOf<DDLItem>()
     private lateinit var adapter: CustomAdapter
     private lateinit var addDDLLauncher: ActivityResultLauncher<Intent>
+    private lateinit var titleBar: TextView
+    private lateinit var excitementText: TextView
+    private lateinit var sharedPreferences: SharedPreferences
+    private lateinit var swipeRefreshLayout: SwipeRefreshLayout
 
     private var pauseRefresh: Boolean = false
 
@@ -46,6 +54,8 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         setContentView(R.layout.activity_main)
 
         DynamicColors.applyToActivitiesIfAvailable(application)
+
+        DynamicColors.applyToActivityIfAvailable(this)
 
         // 获取主题中的 colorSurface 值
         val colorSurface = getThemeColor(android.R.attr.colorBackground)
@@ -59,11 +69,16 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         addEventButton = findViewById(R.id.addEvent)
         settingsButton = findViewById(R.id.settingsButton)
 
-        // 加载数据库中的数据
+        // 设置 RecyclerView
         val itemList = databaseHelper.getAllDDLs()
+        adapter = CustomAdapter(itemList, this)
+        adapter.setSwipeListener(this)
+        recyclerView.layoutManager = LinearLayoutManager(this)
+        recyclerView.adapter = adapter
 
         // 设置 RecyclerView
         adapter = CustomAdapter(itemList, this)
+        adapter.updateData(itemList)
         adapter.setSwipeListener(this)
         // 设置单击监听器
         adapter.setOnItemClickListener(object : CustomAdapter.OnItemClickListener {
@@ -100,14 +115,26 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
                             }
                             1 -> {
                                 // 删除操作
-                                val item = adapter.itemList[position]
-                                databaseHelper.deleteDDL(item.id)
-                                adapter.updateData(databaseHelper.getAllDDLs())
-                                Toast.makeText(this@MainActivity, R.string.toast_deletion, Toast.LENGTH_SHORT).show()
-                                pauseRefresh = false
+                                triggerVibration(this@MainActivity, 200)
+                                MaterialAlertDialogBuilder(this@MainActivity)
+                                    .setTitle(R.string.alert_delete_title)
+                                    .setMessage(R.string.alert_delete_message)
+                                    .setNeutralButton(resources.getString(R.string.cancel)) { dialog, _ ->
+                                        adapter.notifyItemChanged(position) // 取消删除，刷新该项
+                                        pauseRefresh = false
+                                    }
+                                    .setPositiveButton(resources.getString(R.string.accept)) { dialog, _ ->
+                                        val item = adapter.itemList[position]
+                                        databaseHelper.deleteDDL(item.id)
+                                        adapter.updateData(databaseHelper.getAllDDLs())
+                                        Toast.makeText(this@MainActivity, R.string.toast_deletion, Toast.LENGTH_SHORT).show()
+                                        pauseRefresh = false
+                                    }
+                                    .show()
                             }
                             2 -> {
                                 // 标记为完成操作
+                                triggerVibration(this@MainActivity, 100)
                                 val item = adapter.itemList[position]
                                 item.isCompleted = !item.isCompleted
                                 databaseHelper.updateDDL(item)
@@ -135,17 +162,106 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         recyclerView.layoutManager = LinearLayoutManager(this)
         recyclerView.adapter = adapter
 
+        swipeRefreshLayout = findViewById(R.id.swipeRefreshLayout)
+        swipeRefreshLayout.setOnRefreshListener {
+            refreshData()
+        }
+
+        // 添加滑动特效
         // 设置ItemTouchHelper
         val itemTouchHelperCallback = object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.LEFT or ItemTouchHelper.RIGHT) {
+
+            // 定义画笔和图标
+            private val paint = Paint()
+            private val deleteIcon = ContextCompat.getDrawable(this@MainActivity, R.drawable.ic_delete) // 🗑图标资源
+            private val checkIcon = ContextCompat.getDrawable(this@MainActivity, R.drawable.ic_check)   // ✅图标资源
+            private val iconMargin = resources.getDimension(R.dimen.icon_margin).toInt()
+            private val cornerRadius = resources.getDimension(R.dimen.item_corner_radius) // 24dp圆角
+
             override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder): Boolean {
                 return false
             }
 
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 when (direction) {
-                    ItemTouchHelper.LEFT -> adapter.onSwipeLeft(viewHolder.adapterPosition)
-                    ItemTouchHelper.RIGHT -> adapter.onSwipeRight(viewHolder.adapterPosition)
+                    ItemTouchHelper.LEFT -> {
+                        triggerVibration(this@MainActivity, 200)
+                        adapter.onSwipeLeft(viewHolder.adapterPosition)
+                    }
+                    ItemTouchHelper.RIGHT -> {
+                        triggerVibration(this@MainActivity, 100)
+                        adapter.onSwipeRight(viewHolder.adapterPosition)
+                    }
                 }
+            }
+
+            override fun onChildDraw(
+                c: Canvas,
+                recyclerView: RecyclerView,
+                viewHolder: RecyclerView.ViewHolder,
+                dX: Float,
+                dY: Float,
+                actionState: Int,
+                isCurrentlyActive: Boolean
+            ) {
+                val itemView = viewHolder.itemView
+                val itemHeight = itemView.bottom - itemView.top
+
+                if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
+                    val path = Path()
+
+                    // 左滑：绘制低饱和度红色背景和🗑图标
+                    if (dX < 0) {
+                        paint.color = Color.parseColor("#FFEBEE") // 低饱和度红色
+
+                        val background = RectF(
+                            itemView.right + dX,
+                            itemView.top.toFloat(),
+                            itemView.right.toFloat(),
+                            itemView.bottom.toFloat()
+                        )
+
+                        path.addRoundRect(background, cornerRadius, cornerRadius, Path.Direction.CW)
+                        c.drawPath(path, paint)
+
+                        deleteIcon?.let {
+                            val iconTop = itemView.top + (itemHeight - it.intrinsicHeight) / 2
+                            val iconLeft = itemView.right - iconMargin - it.intrinsicWidth
+                            val iconRight = itemView.right - iconMargin
+                            val iconBottom = iconTop + it.intrinsicHeight
+
+                            it.setBounds(iconLeft, iconTop, iconRight, iconBottom)
+                            it.draw(c)
+                        }
+                    }
+
+                    // 右滑：绘制低饱和度绿色背景和✅图标
+                    if (dX > 0) {
+                        paint.color = Color.parseColor("#E8F5E9") // 低饱和度绿色
+
+                        val background = RectF(
+                            itemView.left.toFloat(),
+                            itemView.top.toFloat(),
+                            itemView.left + dX,
+                            itemView.bottom.toFloat()
+                        )
+
+                        path.addRoundRect(background, cornerRadius, cornerRadius, Path.Direction.CW)
+                        c.drawPath(path, paint)
+
+                        checkIcon?.let {
+                            val iconTop = itemView.top + (itemHeight - it.intrinsicHeight) / 2
+                            val iconLeft = itemView.left + iconMargin
+                            val iconRight = itemView.left + iconMargin + it.intrinsicWidth
+                            val iconBottom = iconTop + it.intrinsicHeight
+
+                            it.setBounds(iconLeft, iconTop, iconRight, iconBottom)
+                            it.draw(c)
+                        }
+                    }
+                }
+
+                super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive)
             }
         }
         ItemTouchHelper(itemTouchHelperCallback).attachToRecyclerView(recyclerView)
@@ -163,11 +279,50 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         }
 
         settingsButton.setOnClickListener {
-            Log.d("MainActivity", "Settings triggered")
+//            Log.d("MainActivity", "Settings triggered")
+            val intent = Intent(this, SettingsActivity::class.java)
+            startActivity(intent)
+        }
+
+        titleBar = findViewById(R.id.titleBar)
+        excitementText = findViewById(R.id.excitementText)
+
+        sharedPreferences = getSharedPreferences("app_settings", MODE_PRIVATE)
+
+        // 检查鼓励语句开关状态
+        val isMotivationalQuotesEnabled = sharedPreferences.getBoolean("motivational_quotes", true)
+        updateTitleAndExcitementText(isMotivationalQuotesEnabled)
+    }
+
+    private fun refreshData() {
+        // 在后台线程获取数据，并在主线程更新 UI
+        CoroutineScope(Dispatchers.Main).launch {
+            val newData = withContext(Dispatchers.IO) {
+                databaseHelper.getAllDDLs()
+            }
+            adapter.updateData(newData)
+            swipeRefreshLayout.isRefreshing = false // 停止刷新动画
+        }
+    }
+
+    private fun updateTitleAndExcitementText(isEnabled: Boolean) {
+        if (isEnabled) {
+            Log.d("MainActivity", "Enabled here")
+            titleBar.textSize = 24f // 调整 Deadliner 尺寸
+            excitementText.visibility = TextView.VISIBLE
+
+            // 随机选择数组中的一条语句
+            val excitementArray = resources.getStringArray(R.array.excitement_array)
+            val randomIndex = (excitementArray.indices).random()
+            excitementText.text = excitementArray[randomIndex]
+        } else {
+            titleBar.textSize = 32f // 设置为默认大小
+            excitementText.visibility = TextView.GONE
         }
     }
 
 
+    @Deprecated("Deprecated in Java")
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
 
@@ -206,6 +361,26 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
                 pauseRefresh = false
             }
             .show()
+    }
+
+    /*
+    * 震动效果📳
+    */
+    fun triggerVibration(context: Context, duration: Long = 100) {
+        val isVibrationOn = sharedPreferences.getBoolean("vibration", true)
+        if (!isVibrationOn) {
+            return
+        }
+
+        val vibrator = context.getSystemService(Context.VIBRATOR_SERVICE) as Vibrator
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            // API 26 及以上版本使用 VibrationEffect
+            vibrator.vibrate(VibrationEffect.createOneShot(duration, VibrationEffect.DEFAULT_AMPLITUDE))
+        } else {
+            // API 25 及以下版本使用过时的 vibrate 方法
+            vibrator.vibrate(duration)
+        }
     }
 
     /**
@@ -259,28 +434,10 @@ class MainActivity : AppCompatActivity(), CustomAdapter.SwipeListener {
         return darkness < 0.5
     }
 
-    private val refreshInterval = 5000L
-    private val handler = Handler(Looper.getMainLooper())
-    private val refreshRunnable = object : Runnable {
-        override fun run() {
-            // 刷新数据
-            if (!pauseRefresh) {
-                adapter.updateData(databaseHelper.getAllDDLs())
-                handler.postDelayed(this, refreshInterval)
-            }
-        }
-    }
-
     override fun onResume() {
         super.onResume()
-        // 开始周期性刷新
-        handler.post(refreshRunnable)
-    }
-
-    override fun onPause() {
-        super.onPause()
-        // 停止周期性刷新
-        handler.removeCallbacks(refreshRunnable)
+        val isMotivationalQuotesEnabled = sharedPreferences.getBoolean("motivational_quotes", true)
+        updateTitleAndExcitementText(isMotivationalQuotesEnabled)
     }
 
     companion object {
